refactor #128: MyAccount 페이지 리팩토링
- 컴포넌트 분리
